Added --external-ip option
This is the equivalent of the -externalip option of bitcoind.
It allows the public IP address of your node to be explicitly given,
instead of asking the peer to detect it automatically.
This is useful when running from behind dual WAN connections,
assymetric routing, or any other situation in which you want
inbound connections to come in on a different IP address than outbound.

Signed-off-by: thelazier <thelazier@gmail.com>